#572 split widget model into separate files
