+++ ng2-components/ng2-activiti-form/src/components/widgets/container/container.widget.ts
@@ -16,7 +16,7 @@
  */
 
 import { Component, Input, AfterViewInit } from '@angular/core';
-import { ContainerModel } from './../widget.model';
+import { ContainerModel } from './../core/index';
 
 import { MATERIAL_DESIGN_DIRECTIVES } from 'ng2-alfresco-core';
 import { PRIMITIVE_WIDGET_DIRECTIVES } from './../index';
