+++ ng2-components/ng2-activiti-form/src/components/widgets/tabs/tabs.widget.ts
@@ -17,7 +17,7 @@
 
 import { Component, Input, AfterViewInit } from '@angular/core';
 import { MATERIAL_DESIGN_DIRECTIVES } from 'ng2-alfresco-core';
-import { TabModel } from './../widget.model';
+import { TabModel } from './../core/index';
 import { ContainerWidget } from './../container/container.widget';
 
 declare let __moduleName: string;
